extract write-worker-source
i need it for my tests
